# Performance improvements for gnmsg script